requirements.txt
# Core
pymongo==4.7.2
python-dotenv==1.0.1

# Video / Image Processing
opencv-python==4.9.0.80
Pillow==10.3.0
decord==0.6.0

# Deep Learning
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# CLIP (OpenAI)
git+https://github.com/openai/CLIP.git

# BLIP (Salesforce via HuggingFace)
transformers==4.41.2
accelerate==0.30.1
sentencepiece==0.2.0

# MediaPipe
mediapipe==0.10.14

# Utilities
numpy==1.26.4
tqdm==4.66.4
loguru==0.7.2
